+++ lib/blobstore/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod bitmask;
 pub mod blob;
 mod blobstore;
 pub mod config;
@@ -9,6 +8,8 @@ mod pages;
 mod tracker;
 
 pub use blob::Blob;
+// The bitmask belongs to the Gridstore variant, it is only public for the benchmarks
+pub use blobstore::gridstore::bitmask;
 pub use blobstore::{Blobstore, BlobstoreReader, BlobstoreView};
 
 use crate::error::BlobstoreError;
